Update README.md
This project focuses on building an email spam detection system using Python and machine learning. The dataset is loaded and preprocessed by retaining only the necessary columns for classification (Category and Message). The Category column is encoded into numerical values where "Spam" is represented as 1 and "Not Spam" as 0. The data is then split into training and testing sets (80% for training and 20% for testing). To process the text data, a CountVectorizer is used to transform the messages into numerical feature vectors suitable for machine learning. A Multinomial Naive Bayes model is trained on the processed data, achieving a measurable accuracy score. Finally, a user-interactive function is implemented, allowing users to input a message and classify it as either "SPAM" or "NOT SPAM" based on the trained model.